requirements.txt
python-telegram-bot[job-queue]
google-genai
ccxt
pandas
numpy
python-dotenv
yfinance
